extra manifest and next_require code into separate file
